Hardcoded graphs for demo
